+++ CHANGES.rst
@@ -18,6 +18,8 @@ Unreleased
   maximum lengths.
 - Fixed bug where :meth:`.Reddit.request` could not handle instances of ``BadRequest``\s
   when the JSON data contained only the keys "reason" and "message".
+- Fixed bug where :meth:`.Reddit.request` could not handle instances of ``BadRequest``\s
+  when the response did not contain valid JSON data.
 
 **Deprecated**