+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Change Log
 ==========
 
+PRAW follows `semantic versioning <http://semver.org/>`_.
+
 Unreleased
 ----------